HBX-1908: Externalize logic from 'JdbcMetadataDescriptor#buildMapping(Metadata)' into new class 'org.hibernate.tool.internal.reveng.BinderMapping'
